requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
black==20.8b1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
fsspec==2021.4.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.16
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
pathspec==0.8.1
Pillow==8.2.0
prettytable==2.1.0
protobuf==3.15.8
PTable==0.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.2.10
pytz==2021.1
PyYAML==5.3.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.2
six==1.15.0
sklearn==0.0
smmap==4.0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.1
torch-tb-profiler==0.1.0
torchmetrics==0.2.0
torchvision==0.9.1
tqdm==4.60.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
yarl==1.6.3
